# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rooms": {
      "get": {
        "operationId": "RoomsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The room catalogue.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "List every room ordered by floor and stable room ID",
        "tags": [
          "rooms"
        ]
      },
      "post": {
        "operationId": "RoomsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomMutationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Create a room as an administrator",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{id}": {
      "patch": {
        "operationId": "RoomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Update a room as an administrator using optimistic version",
        "tags": [
          "rooms"
        ]
      },
      "get": {
        "operationId": "RoomsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The room does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Read room details",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{id}/image": {
      "post": {
        "operationId": "RoomsController_uploadImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "image": {
                    "format": "binary",
                    "type": "string"
                  },
                  "version": {
                    "minimum": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "image",
                  "version"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "413": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Upload or replace a PNG/JPEG room image",
        "tags": [
          "rooms"
        ]
      },
      "delete": {
        "operationId": "RoomsController_removeImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageVersionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Remove a room image using optimistic version",
        "tags": [
          "rooms"
        ]
      },
      "get": {
        "operationId": "RoomsController_image",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The room or its image does not exist.",
            "content": {
              "image/png": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              },
              "image/jpeg": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Read the room image when one has been uploaded",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/search": {
      "get": {
        "operationId": "RoomsController_search",
        "parameters": [
          {
            "name": "equipment",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "tv",
                  "projector",
                  "whiteboard",
                  "video_conferencing"
                ]
              }
            }
          },
          {
            "name": "floor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minimumCapacity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Find available rooms free for a future work-time interval",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{id}/schedule": {
      "get": {
        "operationId": "RoomsController_schedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomScheduleDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Read one room schedule for a calendar day",
        "tags": [
          "rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "Roomly API",
    "description": "HTTP API for the Roomly meeting room booking service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Анна Соколова"
          },
          "email": {
            "type": "string",
            "example": "employee@northstar.local"
          },
          "role": {
            "type": "string",
            "enum": [
              "employee",
              "admin"
            ]
          }
        },
        "required": [
          "name",
          "email",
          "role"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Не удалось обработать запрос"
          },
          "fieldErrors": {
            "type": "string",
            "example": [
              "subject: Укажите тему"
            ]
          },
          "requestId": {
            "type": "string",
            "example": "f026e5ca-01e4-4c79-a52e-c44d9c32d967"
          }
        },
        "required": [
          "status",
          "code",
          "message",
          "fieldErrors",
          "requestId"
        ]
      },
      "RoomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "room-atlas"
          },
          "name": {
            "type": "string",
            "example": "Атлас"
          },
          "floor": {
            "type": "number",
            "example": 2
          },
          "location": {
            "type": "string",
            "example": "Крыло A"
          },
          "capacity": {
            "type": "number",
            "example": 12
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Переговорная у окна"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "unavailable"
            ]
          },
          "equipment": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "tv",
                "projector",
                "whiteboard",
                "video_conferencing"
              ]
            }
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "version": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "id",
          "name",
          "floor",
          "location",
          "capacity",
          "status",
          "equipment",
          "version"
        ]
      },
      "RoomMutationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Атлас"
          },
          "floor": {
            "type": "number",
            "example": 2,
            "minimum": 1,
            "maximum": 99
          },
          "location": {
            "type": "string",
            "example": "Крыло A"
          },
          "capacity": {
            "type": "number",
            "example": 12,
            "minimum": 1,
            "maximum": 1000
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Переговорная у окна"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "unavailable"
            ],
            "default": "available"
          },
          "equipment": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "tv",
                "projector",
                "whiteboard",
                "video_conferencing"
              ]
            }
          }
        },
        "required": [
          "name",
          "floor",
          "location",
          "capacity"
        ]
      },
      "RoomUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Атлас"
          },
          "floor": {
            "type": "number",
            "example": 2,
            "minimum": 1,
            "maximum": 99
          },
          "location": {
            "type": "string",
            "example": "Крыло A"
          },
          "capacity": {
            "type": "number",
            "example": 12,
            "minimum": 1,
            "maximum": 1000
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Переговорная у окна"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "unavailable"
            ],
            "default": "available"
          },
          "equipment": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "tv",
                "projector",
                "whiteboard",
                "video_conferencing"
              ]
            }
          },
          "version": {
            "type": "number",
            "example": 1,
            "minimum": 1
          }
        },
        "required": [
          "name",
          "floor",
          "location",
          "capacity",
          "version"
        ]
      },
      "ImageVersionDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "example": 1,
            "minimum": 1
          }
        },
        "required": [
          "version"
        ]
      },
      "ScheduleEntryDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string",
            "example": "10:00"
          },
          "endsAt": {
            "type": "string",
            "example": "11:00"
          },
          "id": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "participants": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "completed"
            ]
          },
          "owner": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "startsAt",
          "endsAt"
        ]
      },
      "RoomScheduleDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-10-05"
          },
          "bookingAllowed": {
            "type": "boolean"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleEntryDto"
            }
          }
        },
        "required": [
          "date",
          "bookingAllowed",
          "entries"
        ]
      },
...
    }
  }
}